+++ docs/Changelog.rst
@@ -3,6 +3,14 @@
 Changelog
 =========
 
+3.4.x (? ? ?)
+-------------
+
+**Bug fixes**
+
+  * Default to trying the cluster state for checking whether indices are closed, and
+    then fall back to using the _cat API (for Amazon ES instances). #519 (untergeek)
+    
 3.4.0 (28 October 2015)
 -----------------------